services/agent-worker/pyproject.toml
[tool.poetry]
name = "mandas-agent-worker"
version = "0.5.0"
description = "Mandas Modular Agent System - Agent Worker (神经系统 - Neural System)"
authors = ["Mandas Team <team@mandas.local>"]
license = "AGPL-3.0"

[tool.poetry.dependencies]
python = "^3.11"
sqlalchemy = "^2.0.23"
asyncpg = "^0.29.0"
redis = "^5.0.1"
pydantic = "^2.5.0"
pydantic-settings = "^2.1.0"
loguru = "^0.7.2"
opentelemetry-api = "^1.21.0"
opentelemetry-sdk = "^1.21.0"
opentelemetry-instrumentation-sqlalchemy = "^0.42b0"
opentelemetry-instrumentation-redis = "^0.42b0"
opentelemetry-exporter-jaeger = "^1.21.0"
httpx = "^0.25.2"
docker = "^6.1.3"
pyautogen = "^0.2.0"
chromadb = "^0.4.18"
sentence-transformers = "^2.2.2"
open-interpreter = "^0.1.16"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-asyncio = "^0.21.1"
black = "^23.11.0"
isort = "^5.12.0"
flake8 = "^6.1.0"
mypy = "^1.7.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py311']

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
